% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoubustIA.R
\name{Uscore}
\alias{Uscore}
\title{Generate U score and its components for each candidate IA timing}
\usage{
Uscore(
  Nmax,
  enroll_time,
  follow_time,
  theta0,
  theta1,
  Pc,
  Pe,
  Pf,
  l,
  a,
  b,
  start,
  end,
  gamma,
  w,
  n_simu = 10000
)
}
\arguments{
\item{Nmax}{Maximum sample size for the trial}

\item{enroll_time}{Estimated duration for patient enrollment}

\item{follow_time}{Estimated maximum follow-up time for each patient}

\item{theta0}{ORR (Overall Response Rate) for external control in the single-arm phase 2 exploratory clinical trial}

\item{theta1}{ORR (Overall Response Rate) for the new treatment}

\item{Pc}{Threshold to reject or not reject the null hypothesis at final analysis}

\item{Pe}{Threshold for Go decision at interim analysis}

\item{Pf}{Threshold for No-Go decision at interim analysis}

\item{l}{Period to consider in the calculation of stability value}

\item{a}{First parameter for Beta prior distribution}

\item{b}{Second parameter for Beta prior distribution}

\item{start}{Start of the candidate IA timing, required to >=2}

\item{end}{End of the candidate IA timing,required to <= Nmax-l}

\item{gamma}{Gamma parameter of Weibull distribution for time to response data in the simulation}

\item{w}{Weight on Go decision at interim analysis, ranging from 0-1}

\item{n_simu}{Number of trials to simulate}
}
\value{
A data frame containing U score and its components (P1k,S1k,P2k,S2k) for each IA timing
}
\description{
Generate U score and its components for each candidate IA timing
}
